network: move bridge option parsing under nmstate
Encapsulate bridge options parsing in network/nmstate because it is used
in production code only in network/nmstate.
