+++ lib/vdsm/network/nmstate/options.py
@@ -1,4 +1,4 @@
-# Copyright 2016-2022 Red Hat, Inc.
+# Copyright 2022 Red Hat, Inc.
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -16,9 +16,6 @@
 #
 # Refer to the README and COPYING files for full details of the license
 #
-
-from __future__ import absolute_import
-from __future__ import division
 
 import os
